Move the remaining testing infrastructure into its own module.
